backend/requirements.txt
asgiref==3.8.1
async-timeout==5.0.1
attrs==24.3.0
autobahn==24.4.2
Automat==24.8.1
cachetools==5.5.0
certifi==2024.8.30
cffi==1.17.0
channels==4.2.0
charset-normalizer==3.3.2
constantly==23.10.4
cryptography==43.0.0
daphne==4.1.2
Django==5.1
django-allauth==64.0.0
django-cors-headers==4.4.0
django-environ==0.11.2
django-formtools==2.5.1
django-otp==1.5.3
django-phonenumber-field==8.0.0
django-two-factor-auth==1.17.0
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
google-auth==2.23.3
hyperlink==21.0.0
idna==3.8
incremental==24.7.2
msgpack==1.1.0
phonenumbers==8.13.44
pillow==10.4.0
pyasn1==0.6.1
pyasn1_modules==0.4.1
pycparser==2.22
PyJWT==2.9.0
pyOpenSSL==24.3.0
pyotp==2.9.0
pypng==0.20220715.0
python-dotenv==1.0.1
qrcode==7.4.2
redis==5.2.1
requests==2.32.3
rsa==4.9
service-identity==24.2.0
sqlparse==0.5.1
tomli==2.2.1
Twisted==24.11.0
txaio==23.1.1
typing_extensions==4.12.2
urllib3==2.2.2
zope.interface==7.2
psycopg[binary]